Revert "2014年7月21日"
This reverts commit 4de3b2b6183871eb04922627842a387ca61ab804.

+++ CFJDS/Common/DataCFSJ.cs
@@ -100,9 +100,9 @@ namespace CFJDS {
         /// <summary>
         /// 违法单价
         /// </summary>
-        private double _illegaUnit;
+        private int _illegaUnit;
 
-        public double IllegaUnit {
+        public int IllegaUnit {
             get { return _illegaUnit; }
             set { _illegaUnit = value; }
         }
